refactor: extract chat config into server/config/chatConfig.js
