+++ server/routes/chat.js
@@ -1,13 +1,13 @@
 // server/routes/chat.js
 const express = require("express");
 const rateLimit = require("express-rate-limit");
-const { GoogleGenerativeAI, GoogleGenerativeAIFetchError } = require("@google/generative-ai");
+const { GoogleGenerativeAI } = require("@google/generative-ai");
 const Product = require("../models/Product");
+const { PRODUCT_KEYWORDS, SYSTEM_PROMPT, getFallbackResponse, PRODUCT_TEMPLATE } = require("../config/chatConfig");
 const router = express.Router();
 
-// Rate Limiter 
 const chatLimiter = rateLimit({
-  windowMs: 15 * 60 * 1000, // 15 minutes
+  windowMs: 15 * 60 * 1000,
   max: 20,
   message: {
     error: "Too many requests, please try again later.",
@@ -15,7 +15,7 @@ const chatLimiter = rateLimit({
   standardHeaders: true,
   legacyHeaders: false,
 });
-// Debug: Check if API key is loaded
+
 console.log("API Key exists:", !!process.env.GEMINI_API_KEY);
 console.log("API Key prefix:", process.env.GEMINI_API_KEY ? process.env.GEMINI_API_KEY.substring(0, 15) + "..." : "MISSING");
 
@@ -26,41 +26,10 @@ if (!process.env.GEMINI_API_KEY) {
 
 const genAI = new GoogleGenerativeAI(process.env.GEMINI_API_KEY);
 
-// Gemini Config
 const modelName = process.env.GEMINI_MODEL_NAME || "gemini-2.5-flash";
 const model = genAI.getGenerativeModel({ model: modelName });
 
-const PRODUCT_KEYWORDS = ["protein", "supplement", "muscle", "gain", "whey", "creatine", "mass"];
-
-const SYSTEM_PROMPT = `You are FitMart's expert fitness assistant.
-Only answer questions related to: workouts, exercise routines, diet, nutrition, 
-protein intake, weight loss, muscle gain, and supplements.
-If the question is unrelated to fitness, politely redirect the user.
-Keep answers concise, practical, and friendly. Use short paragraphs.
-**Use bold text (surround important words with **) to highlight key information like numbers, recommendations, and important terms.**`;
-
-// Enhanced fallback responses with bold formatting
-const getFallbackResponse = (message) => {
-  const lowerMsg = message.toLowerCase();
-
-  if (lowerMsg.includes("protein")) {
-    return "**For optimal protein intake**, aim for **1.6-2.2g per kg** of body weight daily. Good sources include **chicken breast (31g/100g)**, **eggs (6g each)**, **Greek yogurt (10g/100g)**, **lentils (9g/100g)**, and **quality whey protein**. Would you like me to recommend some protein supplements from our store?";
-  }
-  else if (lowerMsg.includes("workout") || lowerMsg.includes("exercise")) {
-    return "**A balanced workout routine** should include: **3-4 strength training sessions** per week focusing on compound movements (**squats, deadlifts, bench press, rows**), plus **2-3 cardio sessions**. Start with **3 sets of 8-12 reps** for each exercise. Remember to **warm up for 5-10 minutes** and **cool down with stretching**!";
-  }
-  else if (lowerMsg.includes("weight loss")) {
-    return "**For sustainable weight loss**: Create a **moderate calorie deficit (300-500 calories below maintenance)**, **prioritize protein intake (1.6-2g per kg body weight)**, combine **strength training with cardio**, get **7-9 hours of sleep**, and **stay hydrated**. Aim for **0.5-1kg loss per week** for healthy results.";
-  }
-  else if (lowerMsg.includes("muscle") || lowerMsg.includes("gain")) {
-    return "**For muscle gain**: Consume a **slight calorie surplus (200-300 above maintenance)**, eat **1.6-2.2g protein per kg body weight**, focus on **progressive overload** in your training, get **adequate sleep (7-9 hours)**, and **stay consistent** with your workouts. **Compound exercises** like **squats, deadlifts, and bench press** are key!";
-  }
-  else {
-    return "I'm here to help with your fitness journey! Feel free to ask about **workouts**, **nutrition**, **protein intake**, **weight loss**, or **muscle gain**. What specific aspect of fitness would you like to know more about?";
-  }
-};
-
-router.post("/", chatLimiter,async (req, res) => {
+router.post("/", chatLimiter, async (req, res) => {
   try {
     const { message } = req.body;
     if (!message?.trim()) {
@@ -96,7 +65,6 @@ router.post("/", chatLimiter,async (req, res) => {
       }
     }
 
-    // Product recommendation logic with bold formatting
     const lower = message.toLowerCase();
     const wantsProduct = PRODUCT_KEYWORDS.some(kw => lower.includes(kw));
 
@@ -110,31 +78,13 @@ router.post("/", chatLimiter,async (req, res) => {
         }).sort({ rating: -1 });
 
         if (product) {
-          // Build product recommendation with bold formatting
-          let productText = "\n\n**💪 Recommended Products**\n";
-          productText += "**" + product.name + "**";
-
-          if (product.brand) {
-            productText += " **by** **" + product.brand + "**";
-          }
-
-          if (product.price) {
-            productText += " **— ₹" + product.price.toLocaleString("en-IN") + "**";
-          }
-
-          if (product.rating) {
-            productText += " **(⭐" + product.rating + "/5)**";
-          }
-
-          reply += productText;
+          reply += PRODUCT_TEMPLATE(product);
         }
       } catch (productError) {
         console.error("Product lookup error:", productError);
-        // Don't fail the whole request if product lookup fails
       }
     }
 
-    // Add note if using fallback
     if (usedFallback) {
       reply += "\n\n*Note: Using enhanced knowledge base. For more detailed responses, ensure API key has available quota.*";
     }
@@ -144,7 +94,7 @@ router.post("/", chatLimiter,async (req, res) => {
     console.error("Chat route error:", err);
     res.status(500).json({
       error: "Failed to generate response",
-      details: process.env.NODE_ENV === 'development' ? err.message : undefined
+      details: process.env.NODE_ENV === "development" ? err.message : undefined,
     });
   }
 });
